1.43.3 (5.3.7f1)
Build-Id: 9e6c8c7d-ef09-43ff-8083-d157bf0136a7
Build-Sha256: a4bc36f9c60481d199baf6fb03b2c9415dbf76713b29aeb412e50c85902bc7e5

+++ Assembly-CSharp/MVWorldObjectClientManagerNetwork.cs
@@ -294,6 +294,11 @@ public class MVWorldObjectClientManagerNetwork : MVWorldObjectClientManager
 	private void DeleteWO(int id)
 	{
 		MVWorldObjectClient worldObjectClient = GetWorldObjectClient(id);
+		if (worldObjectClient == null)
+		{
+			Debug.LogError("Wo not found. This is most likely because non recursive delete server side");
+			return;
+		}
 		worldObjectMapping.RemoveWorldObjectFromTypeSet(worldObjectClient);
 		OnWorldObjectDestroyed(worldObjectClient.Id);
 		worldObjectClient.Destroy();
